Refactored write ahead stuff from streaming.storage to streaming.util
